Slight update. Added comments, added ball texture, decreased temp launch force.
